refactor(pack): refactor xdsl add email
