+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/emailPro/add/add.routing.js
@@ -1,0 +1,11 @@
+import controller from './add.controller';
+import template from './add.html';
+
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state('telecom.packs.pack.emailPro-add', {
+    url: '/xdsl-email/add',
+    template,
+    controller,
+    controllerAs: '$ctrl',
+  });
+};
