--- packages/manager/apps/telecom/src/app/telecom/pack/slots/emailPro/add/pack-emailPro-add.js
@@ -1,11 +1,0 @@
-angular.module('managerApp').config(($stateProvider) => {
-  $stateProvider.state('telecom.packs.pack.emailPro-add', {
-    url: '/xdsl-email/add',
-
-    templateUrl: 'app/telecom/pack/slots/emailPro/add/pack-emailPro-add.html',
-    controller: 'PackEmailProAddCtrl',
-    controllerAs: '$ctrl',
-
-    translations: { value: ['.'], format: 'json' },
-  });
-});
